+++ src/main/java/frc/robot/commands/AutoAimRight.java
@@ -46,7 +46,7 @@ public class AutoAimRight extends CommandBase {
       sysLimeLight.GetBotPose2d(), 
       0, 
       sysVMXPi.vmxPi.getAngle(), 
-      AprilTagCoordinates.AprilTagCoord_Trans2d(Integer.parseInt(sysLimeLight.limeLight_currentlyViewedAprilTag), 3),
+      AprilTagCoordinates.AprilTagCoord_Trans2d(Double.parseDouble(sysLimeLight.limeLight_currentlyViewedAprilTag), 3),
       0,
       sysVMXPi.vmxPi.getAngle()
     );
